Fix the github CI action

+++ ic-stable-structures/src/memory_mapped_files/memory.rs
@@ -29,13 +29,13 @@ impl MemoryMappedFileMemory {
 
 impl Memory for MemoryMappedFileMemory {
     fn size(&self) -> u64 {
-        self.0.borrow().len() / WASM_PAGE_SIZE_IN_BYTES as u64
+        self.0.borrow().len() / WASM_PAGE_SIZE_IN_BYTES
     }
 
     fn grow(&self, pages: u64) -> i64 {
         let mut memory = self.0.borrow_mut();
         let old_size = memory.len();
-        let bytes_to_add = pages * (WASM_PAGE_SIZE_IN_BYTES as u64);
+        let bytes_to_add = pages * WASM_PAGE_SIZE_IN_BYTES;
         let new_length = memory
             .resize(old_size + bytes_to_add)
             .expect("failed to resize memory-mapped file");
@@ -43,7 +43,7 @@ impl Memory for MemoryMappedFileMemory {
             .zero_range(old_size, bytes_to_add)
             .expect("should succeed to zero new memory");
 
-        (new_length / WASM_PAGE_SIZE_IN_BYTES as u64) as i64
+        (new_length / WASM_PAGE_SIZE_IN_BYTES) as i64
     }
 
     fn read(&self, offset: u64, dst: &mut [u8]) {
